v1.2 ROLLBACK
Rolled back to earlier version.

Note: As of now, the application will not deploy as it cannot find the audio files within it. As a result, the application must be run using VS code for now. I have looked into different ways to get the audio files to load, but none have worked unfortunately.